refactor: UseCaseのDI削除
