+++ app/src/main/java/com/example/flush/MainViewModel.kt
@@ -3,6 +3,7 @@ package com.example.flush
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.flush.domain.repository.AuthRepository
+import com.example.flush.domain.use_case.GetCurrentUserUseCase
 import com.example.flush.ui.navigation.Screen
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -14,14 +15,14 @@ import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
-    private val authRepository: AuthRepository,
+    private val getCurrentUserUseCase: GetCurrentUserUseCase
 ) : ViewModel() {
     private val _startDestination = MutableStateFlow<Screen?>(null)
     val startDestination: StateFlow<Screen?> = _startDestination.asStateFlow()
 
     init {
         viewModelScope.launch {
-            val isUserLoggedIn = authRepository.getCurrentUser() != null
+            val isUserLoggedIn = getCurrentUserUseCase() != null
             _startDestination.update {
                 if (isUserLoggedIn) {
                     Screen.Search
