+++ app/src/main/java/com/example/flush/domain/use_case/GetCurrentUserUseCase.kt
@@ -1,0 +1,10 @@
+package com.example.flush.domain.use_case
+
+import com.example.flush.domain.repository.AuthRepository
+import javax.inject.Inject
+
+class GetCurrentUserUseCase @Inject constructor(
+    private val authRepository: AuthRepository,
+) {
+    operator fun invoke() = authRepository.getCurrentUser()
+}
